Fix more mypy errors
