+++ airbyte-cdk/python/airbyte_cdk/sources/declarative/partition_routers/single_partition_router.py
@@ -5,12 +5,12 @@
 from dataclasses import InitVar, dataclass
 from typing import Any, Iterable, Mapping, Optional
 
-from airbyte_cdk.sources.declarative.partition_routers.partition_router import PartitionRouter
+from airbyte_cdk.sources.declarative.stream_slicers.stream_slicer import StreamSlicer
 from airbyte_cdk.sources.types import StreamSlice, StreamState
 
 
 @dataclass
-class SinglePartitionRouter(PartitionRouter):
+class SinglePartitionRouter(StreamSlicer):
     """Partition router returning only a stream slice"""
 
     parameters: InitVar[Mapping[str, Any]]
